Factor out common html settings page
