+++ lib/ui/cardano-wallet-ui.cabal
@@ -55,7 +55,7 @@ library
     Cardano.Wallet.UI.Personal.Html.Pages.About
     Cardano.Wallet.UI.Personal.Html.Pages.Addresses
     Cardano.Wallet.UI.Personal.Html.Pages.Page
-    Cardano.Wallet.UI.Personal.Html.Pages.Settings
+    Cardano.Wallet.UI.Common.Html.Pages.Settings
     Cardano.Wallet.UI.Personal.Html.Pages.Wallet
     Cardano.Wallet.UI.Personal.Html.Pages.Wallets
     Cardano.Wallet.UI.Personal.Html.Pages.Wallets.NewWallet
